Support ARWKV SFT

+++ main/dataset_convert_rlhf_h5.py
@@ -26,7 +26,7 @@ pipeline = PIPELINE(mode='world')
 parser = ArgumentParser()
 
 parser.add_argument("--load_model", default="myfolder/Outputs/rwkv-x070-2b9-cje-instruct-1.pth", type=str)
-parser.add_argument("--model_strategy", default="fp8", type=str)
+parser.add_argument("--model_strategy", default="bf16", type=str)
 parser.add_argument("--rlhf_input_folder", default="myfolder/RLHF/BANCHO", type=str)
 parser.add_argument("--rlhf_output_h5", default="myfolder/RLHF/bancho.h5", type=str)
 parser.add_argument("--endtoken", default="\n\n\x17", type=str)
